# Reduce index memory