@aawsome 2020-06-12
Is now rebased and contains three commits for each individual change.

@MichaelEischer / @rawtaz Thanks for giving the hint about `git rebase -i`! This command rocks!