@aawsome 2020-06-12
There was another issue when saving pointers to the packID array: Every `append` may copy the array. As there are already references to the old array, it will never be garbage collected...
Hence I switched this to saving the index of `packs`.
I did however something wrong when using `git rebase`. So, the commits about saving packs and saving duplicate blobs separately are now merged into one commit - sorry about that!