@aawsome 2020-06-13
> I actually had implemented a proof-of-concept of rather similar changes in https://github.com/MichaelEischer/restic/commits/opt-hash-idx . However, I never got around to clean-up the code. Feel free to "steal" useful parts.

Oh, I wasn't aware of this branch - I did steal some parts, thank you :smile: 

> For deduplicating packIDs to a packIndex, I've also used an additional hashmap during index construction to ensure that each packID is actually only added once. With the `StorePack` method from #2773 that's probably no longer necessary. See [d2086a1](https://github.com/restic/restic/commit/d2086a1bf798721ec5a66699b787cc930f37fd4c) .

I added when using `Store` - maybe `Store` will be removed in future, then this code can also be cleaned. 

> You might want to take the `BenchmarkIndexAlloc` from there. The commit message also contains a rather detailed reasoning how the changes work and why they were made. I can tell from my own experience with e.g. some config file changes from roughly a decade ago which had no explanation at all (and no, the original author didn't remember the why anymore), that it is extremely useful to include a detailed commit message when the reason for the change is not completely obvious.

I added the benchmark (thanks for the hint!) together with a detailed comment of the current state in the first commit.
With the following commits I also changed this comment to make it clear, what the impact of the change is. Also added the explanation why using "inlining" of indexEntry to the `map` in the commit.

I hope all comments are understandable; happy to get feedback or suggestions for improvements!