@aawsome 2020-06-14
Here are the benchmark results on my laptop:

before the changes (only including commit 7419844):
```
BenchmarkPackerManager-4                           	       4	 265876473 ns/op	 198.21 MB/s	   93398 B/op	    1708 allocs/op
BenchmarkDecodeIndex-4                             	   77702	     16464 ns/op
BenchmarkIndexHasUnknown-4                         	13663558	        85.7 ns/op
BenchmarkIndexHasKnown-4                           	13718383	        87.5 ns/op
BenchmarkIndexAlloc-4                              	       1	1531219826 ns/op	904583360 B/op	 2707659 allocs/op
BenchmarkMasterIndexLookupSingleIndex-4            	 6282409	       217 ns/op
BenchmarkMasterIndexLookupMultipleIndex-4          	 1929421	       615 ns/op
BenchmarkMasterIndexLookupSingleIndexUnknown-4     	 8851906	       135 ns/op
BenchmarkMasterIndexLookupMultipleIndexUnknown-4   	 2059674	       564 ns/op
BenchmarkSaveAndEncrypt-4                          	 6194496	       191 ns/op	22009700.59 MB/s	       3 B/op	       0 allocs/op
BenchmarkLoadTree-4                                	    3322	    355004 ns/op
BenchmarkLoadBlob-4                                	     144	   8240706 ns/op	 121.35 MB/s
BenchmarkLoadAndDecrypt-4                          	     126	   9521347 ns/op	 105.03 MB/s
BenchmarkLoadIndex-4                               	      33	  40962411 ns/op
ok  	github.com/restic/restic/internal/repository	165.710s
```

with changes:
```
BenchmarkPackerManager-4                           	       4	 266619089 ns/op	 197.66 MB/s	   93466 B/op	    1709 allocs/op
BenchmarkDecodeIndex-4                             	   75967	     16353 ns/op
BenchmarkIndexHasUnknown-4                         	14869858	        85.5 ns/op
BenchmarkIndexHasKnown-4                           	13833954	        86.4 ns/op
BenchmarkIndexAlloc-4                              	       1	1130858274 ns/op	785641912 B/op	  976445 allocs/op
BenchmarkMasterIndexLookupSingleIndex-4            	 5429026	       223 ns/op
BenchmarkMasterIndexLookupMultipleIndex-4          	 1880902	       630 ns/op
BenchmarkMasterIndexLookupSingleIndexUnknown-4     	 8154114	       135 ns/op
BenchmarkMasterIndexLookupMultipleIndexUnknown-4   	 2070068	       556 ns/op
BenchmarkSaveAndEncrypt-4                          	 6143935	       192 ns/op	21827305.25 MB/s	       3 B/op	       0 allocs/op
BenchmarkLoadTree-4                                	    3522	    357187 ns/op
BenchmarkLoadBlob-4                                	     145	   8296829 ns/op	 120.53 MB/s
BenchmarkLoadAndDecrypt-4                          	     126	   9814544 ns/op	 101.89 MB/s
BenchmarkLoadIndex-4                               	      32	  40884265 ns/op

ok  	github.com/restic/restic/internal/repository	114.982s
```